refactor: move user with profile from onboarding views to global utils
